Work towards updating for no lender ids
This is partial progress unfortunately, i.e., not a working state yet. I
am wondering whether to keep the address route parameter for lenders or
just use the connected address (advantage of the route is that one can
view a lender without being connected as them; but we could also use the
trick of connecting as a mock account for that...).